chore: moved description loading related functions from format_schema to bird_utils
