+++ server/utilities/constants/utilities/format_schema/indexing_constants.py
@@ -8,10 +8,8 @@ COLUMN_EXAMPLES_KEY = "column_examples"
 COLUMN_NAME_KEY = "column_name"
 COLUMN_PRIMARY_KEY = "primary_key"
 COLUMN_TYPE_KEY = "column_type"
-COLUMNS_KEY = "columns"
 FOREIGN_KEY_FROM_COLUMN_KEY = 'from_column'
 FOREIGN_KEY_TO_COLUMN_KEY = 'to_column'
 FOREIGN_KEY_TO_TABLE_KEY = 'to_table'
-TABLE_DESCRIPTION_KEY = "table_description"
 TABLE_FOREIGN_KEY = "foreign_keys"
 TABLE_NAME_KEY = "table_name"
