+++ server/utilities/constants/utilities/format_schema/response_messages.py
@@ -5,5 +5,3 @@ These messages are used to standardize the responses sent to the client, ensurin
 consistency and clarity in communication.
 """
 ERROR_UNSUPPORTED_FORMAT_TYPE = "Unsupported format schema type: {format_type}"
-WARNING_DATABASE_DESCRIPTION_FILE_NOT_FOUND = "Database description file not found: {file_path}"
-WARNING_TABLE_DESCRIPTION_FILE_NOT_FOUND = "Table description file not found: {file_path}"
